Cleaned up and moved Composing to a separate file.
